fix: Refactor code to set data for events and remove events
The code has been refactored to set data for events based on the given object, including updating the title, all-day status, start and end dates/times. It also handles the removal of events based on the provided filter.

The function `setData` now iterates through each object in the `data.objects` array. If an event with the corresponding ID exists in the calendar, it updates the event properties accordingly. If the `filter` object is provided and the `remove` property is true, the event is removed.

If an event with the corresponding ID doesn't exist in the calendar, a new event object is created with the provided data and pushed to the `eventSource` array.

The `removeEvent` function has been updated to receive the `element` parameter instead of the calendar object, as per the TODO comment.